# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CloudsWork Solar Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get plan details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}/purchase:
    post:
      operationId: purchasePlan
      tags: [plans]
      summary: Purchase a solar plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasedPlan"
...
  /profile/update:
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/change-password:
    post:
      operationId: changePassword
      tags: [profile]
      summary: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /profile/banks:
    get:
      operationId: getBankAccounts
      tags: [profile]
      summary: Get user bank accounts
      responses:
        "200":
          description: Bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: addBankAccount
      tags: [profile]
      summary: Add a bank account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "201":
          description: Bank account added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"

  # SUPPORT
...
  /admin/plans:
    get:
      operationId: getAdminPlans
      tags: [admin]
      summary: List all plans (admin)
      responses:
        "200":
          description: Plans list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [admin]
      summary: Create a new plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    put:
      operationId: updatePlan
      tags: [admin]
      summary: Update a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [admin]
      summary: Delete a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        nickname:
          type: string
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        isSuspended:
          type: boolean
        walletBalance:
          type: number
        depositBalance:
          type: number
        withdrawBalance:
          type: number
        createdAt:
          type: string
      required: [id, phone, nickname, referralCode, isAdmin, isSuspended, walletBalance, depositBalance, withdrawBalance, createdAt]
...
    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        price:
          type: number
        dailyProfitRate:
          type: number
        durationDays:
          type: integer
        totalReturn:
          type: number
        maxPurchases:
          type: ["integer", "null"]
        currentPurchases:
          type: integer
        isActive:
          type: boolean
        description:
          type: ["string", "null"]
      required: [id, name, price, dailyProfitRate, durationDays, totalReturn, currentPurchases, isActive]
    PurchasedPlan:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        plan:
          $ref: "#/components/schemas/Plan"
        userId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        totalEarned:
          type: number
        status:
          type: string
          enum: [active, completed, cancelled]
        daysRemaining:
          type: integer
      required: [id, planId, plan, userId, startDate, endDate, totalEarned, status, daysRemaining]
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        nickname:
          type: string
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        referralCode:
          type: string
        totalDeposit:
          type: number
        totalWithdraw:
          type: number
        totalProfit:
          type: number
        totalCommission:
          type: number
        memberSince:
          type: string
      required: [id, phone, nickname, referralCode, totalDeposit, totalWithdraw, totalProfit, totalCommission, memberSince]
    BankAccount:
      type: object
      properties:
        id:
          type: integer
        accountTitle:
          type: string
        iban:
          type: string
        bankName:
          type: string
        isDefault:
          type: boolean
      required: [id, accountTitle, iban, bankName, isDefault]
...
    LoginInput:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required: [phone, password]
    RegisterInput:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        nickname:
          type: string
        referralCode:
          type: ["string", "null"]
      required: [phone, password, confirmPassword, nickname]
    ForgotPasswordInput:
      type: object
      properties:
        phone:
          type: string
      required: [phone]
...
    ProfileInput:
      type: object
      properties:
        nickname:
          type: string
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    ChangePasswordInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
        confirmPassword:
          type: string
      required: [currentPassword, newPassword, confirmPassword]
    BankAccountInput:
      type: object
      properties:
        accountTitle:
          type: string
        iban:
          type: string
        bankName:
          type: string
        isDefault:
          type: boolean
      required: [accountTitle, iban, bankName]
...
    PlanInput:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        price:
          type: number
        dailyProfitRate:
          type: number
        durationDays:
          type: integer
        maxPurchases:
          type: ["integer", "null"]
        isActive:
          type: boolean
        description:
          type: ["string", "null"]
      required: [name, price, dailyProfitRate, durationDays, isActive]
...