# lib/api-zod/src/generated/types/purchasedPlan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CloudsWork Solar Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Plan } from './plan';
import type { PurchasedPlanStatus } from './purchasedPlanStatus';

export interface PurchasedPlan {
  id: number;
  planId: number;
  plan: Plan;
  userId: number;
  startDate: string;
  endDate: string;
  totalEarned: number;
  status: PurchasedPlanStatus;
  daysRemaining: number;
}